Move v1 under mgmt/rest/v1
Implements the API interface to v1.
