+++ mgmt/rest/client/client.go
@@ -38,7 +38,7 @@ import (
 
 	"github.com/asaskevich/govalidator"
 
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/rbody"
 )
 
 var (
