+++ mgmt/rest/client/plugin.go
@@ -27,7 +27,7 @@ import (
 	"time"
 
 	"github.com/intelsdi-x/snap/core/serror"
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/rbody"
 )
 
 // LoadPlugin loads plugins for the given plugin names.
