+++ mgmt/rest/client/task.go
@@ -29,7 +29,7 @@ import (
 	"time"
 
 	"github.com/intelsdi-x/snap/core"
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/rbody"
 	"github.com/intelsdi-x/snap/scheduler/wmap"
 )
