+++ mgmt/rest/server.go
@@ -20,15 +20,11 @@ limitations under the License.
 package rest
 
 import (
-	"bytes"
 	"crypto/tls"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"net"
 	"net/http"
-	"net/http/pprof"
-	"strings"
 	"sync"
 	"time"
 
@@ -36,32 +32,8 @@ import (
 	"github.com/julienschmidt/httprouter"
 	"github.com/urfave/negroni"
 
-	"github.com/intelsdi-x/snap/core"
-	"github.com/intelsdi-x/snap/core/cdata"
-	"github.com/intelsdi-x/snap/core/serror"
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
-	"github.com/intelsdi-x/snap/mgmt/tribe/agreement"
-	cschedule "github.com/intelsdi-x/snap/pkg/schedule"
-	"github.com/intelsdi-x/snap/pkg/stringutils"
-	"github.com/intelsdi-x/snap/scheduler/wmap"
-)
-
-const (
-	APIVersion = 1
-)
-
-// default configuration values
-const (
-	defaultEnable          bool   = true
-	defaultPort            int    = 8181
-	defaultAddress         string = ""
-	defaultHTTPS           bool   = false
-	defaultRestCertificate string = ""
-	defaultRestKey         string = ""
-	defaultAuth            bool   = false
-	defaultAuthPassword    string = ""
-	defaultPortSetByConfig bool   = false
-	defaultPprof           bool   = false
+	"github.com/intelsdi-x/snap/mgmt/rest/api"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1"
 )
 
 var (
@@ -71,111 +43,8 @@ var (
 	protocolPrefix = "http"
 )
 
-// holds the configuration passed in through the SNAP config file
-//   Note: if this struct is modified, then the switch statement in the
-//         UnmarshalJSON method in this same file needs to be modified to
-//         match the field mapping that is defined here
-type Config struct {
-	Enable           bool   `json:"enable"yaml:"enable"`
-	Port             int    `json:"port"yaml:"port"`
-	Address          string `json:"addr"yaml:"addr"`
-	HTTPS            bool   `json:"https"yaml:"https"`
-	RestCertificate  string `json:"rest_certificate"yaml:"rest_certificate"`
-	RestKey          string `json:"rest_key"yaml:"rest_key"`
-	RestAuth         bool   `json:"rest_auth"yaml:"rest_auth"`
-	RestAuthPassword string `json:"rest_auth_password"yaml:"rest_auth_password"`
-	portSetByConfig  bool   ``
-	Pprof            bool   `json:"pprof"yaml:"pprof"`
-}
-
-const (
-	CONFIG_CONSTRAINTS = `
-			"restapi" : {
-				"type": ["object", "null"],
-				"properties" : {
-					"enable": {
-						"type": "boolean"
-					},
-					"https" : {
-						"type": "boolean"
-					},
-					"rest_auth": {
-						"type": "boolean"
-					},
-					"rest_auth_password": {
-						"type": "string"
-					},
-					"rest_certificate": {
-						"type": "string"
-					},
-					"rest_key" : {
-						"type": "string"
-					},
-					"port" : {
-						"type": "integer",
-						"minimum": 1,
-						"maximum": 65535
-					},
-					"addr" : {
-						"type": "string"
-					},
-					"pprof": {
-						"type": "boolean"
-					}
-				},
-				"additionalProperties": false
-			}
-	`
-)
-
-type managesMetrics interface {
-	MetricCatalog() ([]core.CatalogedMetric, error)
-	FetchMetrics(core.Namespace, int) ([]core.CatalogedMetric, error)
-	GetMetricVersions(core.Namespace) ([]core.CatalogedMetric, error)
-	GetMetric(core.Namespace, int) (core.CatalogedMetric, error)
-	Load(*core.RequestedPlugin) (core.CatalogedPlugin, serror.SnapError)
-	Unload(core.Plugin) (core.CatalogedPlugin, serror.SnapError)
-	PluginCatalog() core.PluginCatalog
-	AvailablePlugins() []core.AvailablePlugin
-	GetAutodiscoverPaths() []string
-}
-
-type managesTasks interface {
-	CreateTask(cschedule.Schedule, *wmap.WorkflowMap, bool, ...core.TaskOption) (core.Task, core.TaskErrors)
-	GetTasks() map[string]core.Task
-	GetTask(string) (core.Task, error)
-	StartTask(string) []serror.SnapError
-	StopTask(string) []serror.SnapError
-	RemoveTask(string) error
-	WatchTask(string, core.TaskWatcherHandler) (core.TaskWatcherCloser, error)
-	EnableTask(string) (core.Task, error)
-}
-
-type managesTribe interface {
-	GetAgreement(name string) (*agreement.Agreement, serror.SnapError)
-	GetAgreements() map[string]*agreement.Agreement
-	AddAgreement(name string) serror.SnapError
-	RemoveAgreement(name string) serror.SnapError
-	JoinAgreement(agreementName, memberName string) serror.SnapError
-	LeaveAgreement(agreementName, memberName string) serror.SnapError
-	GetMembers() []string
-	GetMember(name string) *agreement.Member
-}
-
-type managesConfig interface {
-	GetPluginConfigDataNode(core.PluginType, string, int) cdata.ConfigDataNode
-	GetPluginConfigDataNodeAll() cdata.ConfigDataNode
-	MergePluginConfigDataNode(pluginType core.PluginType, name string, ver int, cdn *cdata.ConfigDataNode) cdata.ConfigDataNode
-	MergePluginConfigDataNodeAll(cdn *cdata.ConfigDataNode) cdata.ConfigDataNode
-	DeletePluginConfigDataNodeField(pluginType core.PluginType, name string, ver int, fields ...string) cdata.ConfigDataNode
-	DeletePluginConfigDataNodeFieldAll(fields ...string) cdata.ConfigDataNode
-}
-
 type Server struct {
-	mm         managesMetrics
-	mt         managesTasks
-	tr         managesTribe
-	mc         managesConfig
+	apis       []api.API
 	n          *negroni.Negroni
 	r          *httprouter.Router
 	snapTLS    *snapTLS
@@ -195,26 +64,26 @@ type Server struct {
 // New creates a REST API server with a given config
 func New(cfg *Config) (*Server, error) {
 	// pull a few parameters from the configuration passed in by snapteld
-	https := cfg.HTTPS
-	cpath := cfg.RestCertificate
-	kpath := cfg.RestKey
-	pprof := cfg.Pprof
 	s := &Server{
 		err:        make(chan error),
 		killChan:   make(chan struct{}),
 		addrString: cfg.Address,
-		pprof:      pprof,
+		pprof:      cfg.Pprof,
 	}
-	if https {
+	if cfg.HTTPS {
 		var err error
-		s.snapTLS, err = newtls(cpath, kpath)
+		s.snapTLS, err = newtls(cfg.RestCertificate, cfg.RestKey)
 		if err != nil {
 			return nil, err
 		}
 		protocolPrefix = "https"
 	}
+	restLogger.Info(fmt.Sprintf("Configuring REST API with HTTPS set to: %v", cfg.HTTPS))
 
-	restLogger.Info(fmt.Sprintf("Configuring REST API with HTTPS set to: %v", https))
+	s.apis = []api.API{
+		v1.New(&s.wg, s.killChan, protocolPrefix),
+	}
+
 	s.n = negroni.New(
 		NewLogger(),
 		negroni.NewRecovery(),
@@ -226,84 +95,28 @@ func New(cfg *Config) (*Server, error) {
 	return s, nil
 }
 
-// GetDefaultConfig gets the default snapteld configuration
-func GetDefaultConfig() *Config {
-	return &Config{
-		Enable:           defaultEnable,
-		Port:             defaultPort,
-		Address:          defaultAddress,
-		HTTPS:            defaultHTTPS,
-		RestCertificate:  defaultRestCertificate,
-		RestKey:          defaultRestKey,
-		RestAuth:         defaultAuth,
-		RestAuthPassword: defaultAuthPassword,
-		portSetByConfig:  defaultPortSetByConfig,
-		Pprof:            defaultPprof,
+func (s *Server) BindMetricManager(m api.Metrics) {
+	for _, apiInstance := range s.apis {
+		apiInstance.BindMetricManager(m)
 	}
 }
 
-// define a method that can be used to determine if the port the RESTful
-// API is listening on was set in the configuration file
-func (c *Config) PortSetByConfigFile() bool {
-	return c.portSetByConfig
+func (s *Server) BindTaskManager(t api.Tasks) {
+	for _, apiInstance := range s.apis {
+		apiInstance.BindTaskManager(t)
+	}
 }
 
-// UnmarshalJSON unmarshals valid json into a Config.  An example Config can be found
-// at github.com/intelsdi-x/snap/blob/master/examples/configs/snap-config-sample.json
-func (c *Config) UnmarshalJSON(data []byte) error {
-	// construct a map of strings to json.RawMessages (to defer the parsing of individual
-	// fields from the unmarshalled interface until later) and unmarshal the input
-	// byte array into that map
-	t := make(map[string]json.RawMessage)
-	if err := json.Unmarshal(data, &t); err != nil {
-		return err
+func (s *Server) BindTribeManager(t api.Tribe) {
+	for _, apiInstance := range s.apis {
+		apiInstance.BindTribeManager(t)
 	}
-	// loop through the individual map elements, parse each in turn, and set
-	// the appropriate field in this configuration
-	for k, v := range t {
-		switch k {
-		case "enable":
-			if err := json.Unmarshal(v, &(c.Enable)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::enable')", err)
-			}
-		case "port":
-			if err := json.Unmarshal(v, &(c.Port)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::port')", err)
-			}
-			c.portSetByConfig = true
-		case "addr":
-			if err := json.Unmarshal(v, &(c.Address)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::addr')", err)
-			}
-		case "https":
-			if err := json.Unmarshal(v, &(c.HTTPS)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::https')", err)
-			}
-		case "rest_certificate":
-			if err := json.Unmarshal(v, &(c.RestCertificate)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::rest_certificate')", err)
-			}
-		case "rest_key":
-			if err := json.Unmarshal(v, &(c.RestKey)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::rest_key')", err)
-			}
-		case "rest_auth":
-			if err := json.Unmarshal(v, &(c.RestAuth)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::rest_auth')", err)
-			}
-		case "rest_auth_password":
-			if err := json.Unmarshal(v, &(c.RestAuthPassword)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::rest_auth_password')", err)
-			}
-		case "pprof":
-			if err := json.Unmarshal(v, &(c.Pprof)); err != nil {
-				return fmt.Errorf("%v (while parsing 'restapi::pprof')", err)
-			}
-		default:
-			return fmt.Errorf("Unrecognized key '%v' in global config file while parsing 'restapi'", k)
-		}
+}
+
+func (s *Server) BindConfigManager(c api.Config) {
+	for _, apiInstance := range s.apis {
+		apiInstance.BindConfigManager(c)
 	}
-	return nil
 }
 
 // SetAPIAuth sets API authentication to enabled or disabled
@@ -411,8 +224,8 @@ func (s *Server) serveTLS(ln net.Listener) {
 	if err != nil {
 		select {
 		case <-s.closingChan:
-			// If we called Stop() then there will be a value in s.closingChan, so
-			// we'll get here and we can exit without showing the error.
+		// If we called Stop() then there will be a value in s.closingChan, so
+		// we'll get here and we can exit without showing the error.
 		default:
 			restLogger.Error(err)
 			s.err <- err
@@ -426,13 +239,22 @@ func (s *Server) serve(ln net.Listener) {
 	if err != nil {
 		select {
 		case <-s.closingChan:
-			// If we called Stop() then there will be a value in s.closingChan, so
-			// we'll get here and we can exit without showing the error.
+		// If we called Stop() then there will be a value in s.closingChan, so
+		// we'll get here and we can exit without showing the error.
 		default:
 			restLogger.Error(err)
 			s.err <- err
 		}
 	}
+}
+
+func (s *Server) addRoutes() {
+	for _, apiInstance := range s.apis {
+		for _, route := range apiInstance.GetRoutes() {
+			s.r.Handle(route.Method, route.Path, route.Handle)
+		}
+	}
+	s.addPprofRoutes()
 }
 
 // Monkey patch ListenAndServe and TCP alive code from https://golang.org/src/net/http/server.go
@@ -451,143 +273,4 @@ func (ln tcpKeepAliveListener) Accept() (c net.Conn, err error) {
 	tc.SetKeepAlive(true)
 	tc.SetKeepAlivePeriod(3 * time.Minute)
 	return tc, nil
-}
-
-func (s *Server) BindMetricManager(m managesMetrics) {
-	s.mm = m
-}
-
-func (s *Server) BindTaskManager(t managesTasks) {
-	s.mt = t
-}
-
-func (s *Server) BindTribeManager(t managesTribe) {
-	s.tr = t
-}
-
-func (s *Server) BindConfigManager(c managesConfig) {
-	s.mc = c
-}
-
-func (s *Server) addRoutes() {
-
-	// plugin routes
-	s.r.GET("/v1/plugins", s.getPlugins)
-	s.r.GET("/v1/plugins/:type", s.getPlugins)
-	s.r.GET("/v1/plugins/:type/:name", s.getPlugins)
-	s.r.GET("/v1/plugins/:type/:name/:version", s.getPlugin)
-	s.r.POST("/v1/plugins", s.loadPlugin)
-	s.r.DELETE("/v1/plugins/:type/:name/:version", s.unloadPlugin)
-	s.r.GET("/v1/plugins/:type/:name/:version/config", s.getPluginConfigItem)
-	s.r.PUT("/v1/plugins/:type/:name/:version/config", s.setPluginConfigItem)
-	s.r.DELETE("/v1/plugins/:type/:name/:version/config", s.deletePluginConfigItem)
-
-	// metric routes
-	s.r.GET("/v1/metrics", s.getMetrics)
-	s.r.GET("/v2/metrics", s.getMetrics)
-
-	s.r.GET("/v1/metrics/*namespace", s.getMetricsFromTree)
-	s.r.GET("/v2/metrics/*namespace", s.getMetricsFromTree)
-
-	// task routes
-	s.r.GET("/v1/tasks", s.getTasks)
-	s.r.GET("/v2/tasks", s.getTasks)
-
-	s.r.GET("/v1/tasks/:id", s.getTask)
-	s.r.GET("/v1/tasks/:id/watch", s.watchTask)
-	s.r.POST("/v1/tasks", s.addTask)
-	s.r.PUT("/v1/tasks/:id/start", s.startTask)
-	s.r.PUT("/v1/tasks/:id/stop", s.stopTask)
-	s.r.DELETE("/v1/tasks/:id", s.removeTask)
-	s.r.PUT("/v1/tasks/:id/enable", s.enableTask)
-
-	// tribe routes
-	if s.tr != nil {
-		s.r.GET("/v1/tribe/agreements", s.getAgreements)
-		s.r.GET("/v2/tribes/agreements", s.getAgreements)
-
-		s.r.POST("/v1/tribe/agreements", s.addAgreement)
-		s.r.POST("/v2/tribes/agreements", s.addAgreement)
-
-		s.r.GET("/v1/tribe/agreements/:name", s.getAgreement)
-		s.r.GET("/v2/tribes/agreements/:name", s.getAgreement)
-
-		s.r.DELETE("/v1/tribe/agreements/:name", s.deleteAgreement)
-		s.r.DELETE("/v2/tribes/agreements/:name", s.deleteAgreement)
-
-		s.r.PUT("/v1/tribe/agreements/:name/join", s.joinAgreement)
-		s.r.PUT("/v2/tribes/agreements/:name/join", s.joinAgreement)
-
-		s.r.DELETE("/v1/tribe/agreements/:name/leave", s.leaveAgreement)
-		s.r.DELETE("/v2/tribes/agreements/:name/leave", s.leaveAgreement)
-
-		s.r.GET("/v1/tribe/members", s.getMembers)
-		s.r.GET("/v2/tribes/members", s.getMembers)
-
-		s.r.GET("/v1/tribe/member/:name", s.getMember)
-		s.r.GET("/v2/tribes/members/:name", s.getMember)
-	}
-
-	// profiling tools routes
-	if s.pprof {
-		s.r.GET("/debug/pprof/", s.index)
-		s.r.GET("/debug/pprof/block", s.index)
-		s.r.GET("/debug/pprof/goroutine", s.index)
-		s.r.GET("/debug/pprof/heap", s.index)
-		s.r.GET("/debug/pprof/threadcreate", s.index)
-		s.r.GET("/debug/pprof/cmdline", s.cmdline)
-		s.r.GET("/debug/pprof/profile", s.profile)
-		s.r.GET("/debug/pprof/symbol", s.symbol)
-		s.r.GET("/debug/pprof/trace", s.trace)
-	}
-}
-
-// profiling tools handlers
-
-func (s *Server) index(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	pprof.Index(w, r)
-}
-
-func (s *Server) cmdline(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	pprof.Cmdline(w, r)
-}
-
-func (s *Server) profile(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	pprof.Profile(w, r)
-}
-
-func (s *Server) symbol(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	pprof.Symbol(w, r)
-}
-
-func (s *Server) trace(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	pprof.Trace(w, r)
-}
-
-func respond(code int, b rbody.Body, w http.ResponseWriter) {
-	resp := &rbody.APIResponse{
-		Meta: &rbody.APIResponseMeta{
-			Code:    code,
-			Message: b.ResponseBodyMessage(),
-			Type:    b.ResponseBodyType(),
-			Version: APIVersion,
-		},
-		Body: b,
-	}
-	if !w.(negroni.ResponseWriter).Written() {
-		w.WriteHeader(code)
-	}
-
-	j, err := json.MarshalIndent(resp, "", "  ")
-	if err != nil {
-		panic(err)
-	}
-	j = bytes.Replace(j, []byte("\\u0026"), []byte("&"), -1)
-	fmt.Fprint(w, string(j))
-}
-
-func parseNamespace(ns string) []string {
-	fc := stringutils.GetFirstChar(ns)
-	ns = strings.Trim(ns, fc)
-	return strings.Split(ns, fc)
 }
