+++ mgmt/rest/v1/fixtures/mock_metric_manager.go
