+++ mgmt/rest/v1/metric.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package rest
+package v1
 
 import (
 	"fmt"
@@ -27,24 +27,19 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/julienschmidt/httprouter"
-
 	"github.com/intelsdi-x/snap/core"
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
-	apiV2 "github.com/intelsdi-x/snap/mgmt/rest/v2"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/rbody"
 	"github.com/intelsdi-x/snap/pkg/stringutils"
+	"github.com/julienschmidt/httprouter"
 )
 
-func (s *Server) getMetrics(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+func (s *apiV1) getMetrics(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
 	ver := 0 // 0: get all metrics
 
 	// If we are provided a parameter with the name 'ns' we need to
 	// perform a query
 	q := r.URL.Query()
 	v := q.Get("ver")
-	// apiVersion should be "v1" or "v2".
-	// If apiVersion != "v2" calls default to V1 functionality
-	apiVersion := r.URL.Path[1:3]
 	ns_query := q.Get("ns")
 	if ns_query != "" {
 		ver = 0 // 0: get all versions
@@ -52,7 +47,7 @@ func (s *Server) getMetrics(w http.ResponseWriter, r *http.Request, _ httprouter
 			var err error
 			ver, err = strconv.Atoi(v)
 			if err != nil {
-				respond(400, rbody.FromError(err), w)
+				rbody.Write(400, rbody.FromError(err), w)
 				return
 			}
 		}
@@ -63,24 +58,24 @@ func (s *Server) getMetrics(w http.ResponseWriter, r *http.Request, _ httprouter
 			ns = ns[:len(ns)-1]
 		}
 
-		mts, err := s.mm.FetchMetrics(core.NewNamespace(ns...), ver)
+		mts, err := s.metricManager.FetchMetrics(core.NewNamespace(ns...), ver)
 		if err != nil {
-			respond(404, rbody.FromError(err), w)
+			rbody.Write(404, rbody.FromError(err), w)
 			return
 		}
-		respondWithMetrics(r.Host, mts, w, apiVersion)
+		respondWithMetrics(r.Host, mts, w)
 		return
 	}
 
-	mts, err := s.mm.MetricCatalog()
+	mts, err := s.metricManager.MetricCatalog()
 	if err != nil {
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
-	respondWithMetrics(r.Host, mts, w, apiVersion)
+	respondWithMetrics(r.Host, mts, w)
 }
 
-func (s *Server) getMetricsFromTree(w http.ResponseWriter, r *http.Request, params httprouter.Params) {
+func (s *apiV1) getMetricsFromTree(w http.ResponseWriter, r *http.Request, params httprouter.Params) {
 	namespace := params.ByName("namespace")
 
 	// we land here if the request contains a trailing slash, because it matches the tree
@@ -100,9 +95,6 @@ func (s *Server) getMetricsFromTree(w http.ResponseWriter, r *http.Request, para
 	)
 	q := r.URL.Query()
 	v := q.Get("ver")
-	// apiVersion should be "v1" or "v2".
-	// If apiVersion != "v2" calls default to V1 functionality
-	apiVersion := r.URL.Path[1:3]
 
 	if ns[len(ns)-1] == "*" {
 		if v == "" {
@@ -110,40 +102,40 @@ func (s *Server) getMetricsFromTree(w http.ResponseWriter, r *http.Request, para
 		} else {
 			ver, err = strconv.Atoi(v)
 			if err != nil {
-				respond(400, rbody.FromError(err), w)
+				rbody.Write(400, rbody.FromError(err), w)
 				return
 			}
 		}
 
-		mts, err := s.mm.FetchMetrics(core.NewNamespace(ns[:len(ns)-1]...), ver)
+		mts, err := s.metricManager.FetchMetrics(core.NewNamespace(ns[:len(ns)-1]...), ver)
 		if err != nil {
-			respond(404, rbody.FromError(err), w)
+			rbody.Write(404, rbody.FromError(err), w)
 			return
 		}
-		respondWithMetrics(r.Host, mts, w, apiVersion)
+		respondWithMetrics(r.Host, mts, w)
 		return
 	}
 
 	// If no version was given, get all version that fall at this namespace.
 	if v == "" {
-		mts, err := s.mm.FetchMetrics(core.NewNamespace(ns...), 0)
+		mts, err := s.metricManager.FetchMetrics(core.NewNamespace(ns...), 0)
 		if err != nil {
-			respond(404, rbody.FromError(err), w)
+			rbody.Write(404, rbody.FromError(err), w)
 			return
 		}
-		respondWithMetrics(r.Host, mts, w, apiVersion)
+		respondWithMetrics(r.Host, mts, w)
 		return
 	}
 
 	// if an explicit version is given, get that single one.
 	ver, err = strconv.Atoi(v)
 	if err != nil {
-		respond(400, rbody.FromError(err), w)
+		rbody.Write(400, rbody.FromError(err), w)
 		return
 	}
-	mt, err := s.mm.GetMetric(core.NewNamespace(ns...), ver)
+	mt, err := s.metricManager.GetMetric(core.NewNamespace(ns...), ver)
 	if err != nil {
-		respond(404, rbody.FromError(err), w)
+		rbody.Write(404, rbody.FromError(err), w)
 		return
 	}
 
@@ -161,19 +153,15 @@ func (s *Server) getMetricsFromTree(w http.ResponseWriter, r *http.Request, para
 		Description:     mt.Description(),
 		Unit:            mt.Unit(),
 		LastAdvertisedTimestamp: mt.LastAdvertisedTime().Unix(),
-		Href: catalogedMetricURI(r.Host, mt),
+		Href: catalogedMetricURI(r.Host, version, mt),
 	}
 	policies := rbody.PolicyTableSlice(mt.Policy().RulesAsTable())
 	mb.Policy = policies
 	b.Metric = mb
-	respond(200, b, w)
+	rbody.Write(200, b, w)
 }
 
-func respondWithMetrics(host string, mts []core.CatalogedMetric, w http.ResponseWriter, version string) {
-	if version == "v2" {
-		apiV2.RespondWithMetrics(host, mts, w)
-		return
-	}
+func respondWithMetrics(host string, mts []core.CatalogedMetric, w http.ResponseWriter) {
 	b := rbody.NewMetricsReturned()
 	for _, m := range mts {
 		policies := rbody.PolicyTableSlice(m.Policy().RulesAsTable())
@@ -191,15 +179,11 @@ func respondWithMetrics(host string, mts []core.CatalogedMetric, w http.Response
 			DynamicElements:         dynamicElements,
 			Unit:                    m.Unit(),
 			Policy:                  policies,
-			Href:                    catalogedMetricURI(host, m),
+			Href:                    catalogedMetricURI(host, version, m),
 		})
 	}
 	sort.Sort(b)
-	respond(200, b, w)
-}
-
-func catalogedMetricURI(host string, mt core.CatalogedMetric) string {
-	return fmt.Sprintf("%s://%s/v1/metrics?ns=%s&ver=%d", protocolPrefix, host, url.QueryEscape(mt.Namespace().String()), mt.Version())
+	rbody.Write(200, b, w)
 }
 
 func getDynamicElements(ns core.Namespace, indexes []int) []rbody.DynamicElement {
@@ -213,4 +197,14 @@ func getDynamicElements(ns core.Namespace, indexes []int) []rbody.DynamicElement
 		})
 	}
 	return elements
+}
+
+func catalogedMetricURI(host, version string, mt core.CatalogedMetric) string {
+	return fmt.Sprintf("%s://%s/%s/metrics?ns=%s&ver=%d", protocolPrefix, host, version, url.QueryEscape(mt.Namespace().String()), mt.Version())
+}
+
+func parseNamespace(ns string) []string {
+	fc := stringutils.GetFirstChar(ns)
+	ns = strings.Trim(ns, fc)
+	return strings.Split(ns, fc)
 }
