+++ mgmt/rest/v1/plugin_test.go
@@ -1,4 +1,4 @@
-// +build legacy
+// +build medium
 
 /*
 http://www.apache.org/licenses/LICENSE-2.0.txt
@@ -19,12 +19,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package rest
+package v1
 
 import (
 	"testing"
 
-	"github.com/intelsdi-x/snap/mgmt/rest/fixtures"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/fixtures"
 	. "github.com/smartystreets/goconvey/convey"
 )
