+++ mgmt/rest/v1/task.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package rest
+package v1
 
 import (
 	"errors"
@@ -28,11 +28,9 @@ import (
 	"time"
 
 	log "github.com/Sirupsen/logrus"
-	apiV2 "github.com/intelsdi-x/snap/mgmt/rest/v2"
-	"github.com/julienschmidt/httprouter"
-
 	"github.com/intelsdi-x/snap/core"
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/rbody"
+	"github.com/julienschmidt/httprouter"
 )
 
 var (
@@ -42,28 +40,23 @@ var (
 	ErrStreamingUnsupported    = errors.New("Streaming unsupported")
 	ErrTaskNotFound            = errors.New("Task not found")
 	ErrTaskDisabledNotRunnable = errors.New("Task is disabled. Cannot be started")
+	ErrNoActionSpecified       = errors.New("No action was specified in the request")
+	ErrWrongAction             = errors.New("Wrong action requested")
 )
 
-func (s *Server) addTask(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	task, err := core.CreateTaskFromContent(r.Body, nil, s.mt.CreateTask)
+func (s *apiV1) addTask(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+	task, err := core.CreateTaskFromContent(r.Body, nil, s.taskManager.CreateTask)
 	if err != nil {
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
 	taskB := rbody.AddSchedulerTaskFromTask(task)
-	taskB.Href = taskURI(r.Host, task)
-	respond(201, taskB, w)
+	taskB.Href = taskURI(r.Host, version, task)
+	rbody.Write(201, taskB, w)
 }
 
-func (s *Server) getTasks(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
-	sts := s.mt.GetTasks()
-	// apiVersion should be "v1" or "v2".
-	// If apiVersion != "v2" calls default to V1 functionality
-	apiVersion := r.URL.Path[1:3]
-	if apiVersion == "v2" {
-		apiV2.GetTasks(w, r, sts)
-		return
-	}
+func (s *apiV1) getTasks(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+	sts := s.taskManager.GetTasks()
 
 	tasks := &rbody.ScheduledTaskListReturned{}
 	tasks.ScheduledTasks = make([]rbody.ScheduledTask, len(sts))
@@ -71,27 +64,27 @@ func (s *Server) getTasks(w http.ResponseWriter, r *http.Request, _ httprouter.P
 	i := 0
 	for _, t := range sts {
 		tasks.ScheduledTasks[i] = *rbody.SchedulerTaskFromTask(t)
-		tasks.ScheduledTasks[i].Href = taskURI(r.Host, t)
+		tasks.ScheduledTasks[i].Href = taskURI(r.Host, version, t)
 		i++
 	}
 	sort.Sort(tasks)
-	respond(200, tasks, w)
+	rbody.Write(200, tasks, w)
 }
 
-func (s *Server) getTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) getTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	id := p.ByName("id")
-	t, err1 := s.mt.GetTask(id)
+	t, err1 := s.taskManager.GetTask(id)
 	if err1 != nil {
-		respond(404, rbody.FromError(err1), w)
+		rbody.Write(404, rbody.FromError(err1), w)
 		return
 	}
 	task := &rbody.ScheduledTaskReturned{}
 	task.AddScheduledTask = *rbody.AddSchedulerTaskFromTask(t)
-	task.Href = taskURI(r.Host, t)
-	respond(200, task, w)
+	task.Href = taskURI(r.Host, version, t)
+	rbody.Write(200, task, w)
 }
 
-func (s *Server) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	s.wg.Add(1)
 	defer s.wg.Done()
 	logger := log.WithFields(log.Fields{
@@ -109,13 +102,13 @@ func (s *Server) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.
 		alive: true,
 		mChan: make(chan rbody.StreamedTaskEvent),
 	}
-	tc, err1 := s.mt.WatchTask(id, tw)
+	tc, err1 := s.taskManager.WatchTask(id, tw)
 	if err1 != nil {
 		if strings.Contains(err1.Error(), ErrTaskNotFound.Error()) {
-			respond(404, rbody.FromError(err1), w)
+			rbody.Write(404, rbody.FromError(err1), w)
 			return
 		}
-		respond(500, rbody.FromError(err1), w)
+		rbody.Write(500, rbody.FromError(err1), w)
 		return
 	}
 
@@ -129,7 +122,7 @@ func (s *Server) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.
 	flusher, ok := w.(http.Flusher)
 	if !ok {
 		// This only works on ResponseWriters that support streaming
-		respond(500, rbody.FromError(ErrStreamingUnsupported), w)
+		rbody.Write(500, rbody.FromError(ErrStreamingUnsupported), w)
 		return
 	}
 	// send initial stream open event
@@ -164,7 +157,7 @@ func (s *Server) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.
 				// Close out watcher removing it from the scheduler
 				tc.Close()
 				// exit since this client is no longer listening
-				respond(200, &rbody.ScheduledTaskWatchingEnded{}, w)
+				rbody.Write(200, &rbody.ScheduledTaskWatchingEnded{}, w)
 			}
 			// If we are at least above our minimum buffer time we flush to send
 			if time.Now().Sub(t).Seconds() > StreamingBufferWindow {
@@ -180,7 +173,7 @@ func (s *Server) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.
 			// Close out watcher removing it from the scheduler
 			tc.Close()
 			// exit since this client is no longer listening
-			respond(200, &rbody.ScheduledTaskWatchingEnded{}, w)
+			rbody.Write(200, &rbody.ScheduledTaskWatchingEnded{}, w)
 			return
 		case <-s.killChan:
 			logger.WithFields(log.Fields{
@@ -191,74 +184,74 @@ func (s *Server) watchTask(w http.ResponseWriter, r *http.Request, p httprouter.
 			// Close out watcher removing it from the scheduler
 			tc.Close()
 			// exit since this client is no longer listening
-			respond(200, &rbody.ScheduledTaskWatchingEnded{}, w)
+			rbody.Write(200, &rbody.ScheduledTaskWatchingEnded{}, w)
 			return
 		}
 	}
 }
 
-func (s *Server) startTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) startTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	id := p.ByName("id")
-	errs := s.mt.StartTask(id)
+	errs := s.taskManager.StartTask(id)
 	if errs != nil {
 		if strings.Contains(errs[0].Error(), ErrTaskNotFound.Error()) {
-			respond(404, rbody.FromSnapErrors(errs), w)
+			rbody.Write(404, rbody.FromSnapErrors(errs), w)
 			return
 		}
 		if strings.Contains(errs[0].Error(), ErrTaskDisabledNotRunnable.Error()) {
-			respond(409, rbody.FromSnapErrors(errs), w)
+			rbody.Write(409, rbody.FromSnapErrors(errs), w)
 			return
 		}
-		respond(500, rbody.FromSnapErrors(errs), w)
+		rbody.Write(500, rbody.FromSnapErrors(errs), w)
 		return
 	}
 	// TODO should return resource
-	respond(200, &rbody.ScheduledTaskStarted{ID: id}, w)
+	rbody.Write(200, &rbody.ScheduledTaskStarted{ID: id}, w)
 }
 
-func (s *Server) stopTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) stopTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	id := p.ByName("id")
-	errs := s.mt.StopTask(id)
+	errs := s.taskManager.StopTask(id)
 	if errs != nil {
 		if strings.Contains(errs[0].Error(), ErrTaskNotFound.Error()) {
-			respond(404, rbody.FromSnapErrors(errs), w)
+			rbody.Write(404, rbody.FromSnapErrors(errs), w)
 			return
 		}
-		respond(500, rbody.FromSnapErrors(errs), w)
+		rbody.Write(500, rbody.FromSnapErrors(errs), w)
 		return
 	}
-	respond(200, &rbody.ScheduledTaskStopped{ID: id}, w)
+	rbody.Write(200, &rbody.ScheduledTaskStopped{ID: id}, w)
 }
 
-func (s *Server) removeTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) removeTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	id := p.ByName("id")
-	err := s.mt.RemoveTask(id)
+	err := s.taskManager.RemoveTask(id)
 	if err != nil {
 		if strings.Contains(err.Error(), ErrTaskNotFound.Error()) {
-			respond(404, rbody.FromError(err), w)
+			rbody.Write(404, rbody.FromError(err), w)
 			return
 		}
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
-	respond(200, &rbody.ScheduledTaskRemoved{ID: id}, w)
+	rbody.Write(200, &rbody.ScheduledTaskRemoved{ID: id}, w)
 }
 
 //enableTask changes the task state from Disabled to Stopped
-func (s *Server) enableTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) enableTask(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	id := p.ByName("id")
-	tsk, err := s.mt.EnableTask(id)
+	tsk, err := s.taskManager.EnableTask(id)
 	if err != nil {
 		if strings.Contains(err.Error(), ErrTaskNotFound.Error()) {
-			respond(404, rbody.FromError(err), w)
+			rbody.Write(404, rbody.FromError(err), w)
 			return
 		}
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
 	task := &rbody.ScheduledTaskEnabled{}
 	task.AddScheduledTask = *rbody.AddSchedulerTaskFromTask(tsk)
-	respond(200, task, w)
+	rbody.Write(200, task, w)
 }
 
 type TaskWatchHandler struct {
@@ -303,6 +296,6 @@ func (t *TaskWatchHandler) CatchTaskDisabled(why string) {
 	}
 }
 
-func taskURI(host string, t core.Task) string {
-	return fmt.Sprintf("%s://%s/v1/tasks/%s", protocolPrefix, host, t.ID())
+func taskURI(host, version string, t core.Task) string {
+	return fmt.Sprintf("%s://%s/%s/tasks/%s", protocolPrefix, host, version, t.ID())
 }
