+++ mgmt/rest/v1/tribe.go
@@ -17,7 +17,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package rest
+package v1
 
 import (
 	"encoding/json"
@@ -28,7 +28,7 @@ import (
 	log "github.com/Sirupsen/logrus"
 
 	"github.com/intelsdi-x/snap/core/serror"
-	"github.com/intelsdi-x/snap/mgmt/rest/rbody"
+	"github.com/intelsdi-x/snap/mgmt/rest/v1/rbody"
 	"github.com/julienschmidt/httprouter"
 )
 
@@ -42,75 +42,75 @@ var (
 	ErrMemberNotFound        = errors.New("Member not found")
 )
 
-func (s *Server) getAgreements(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
+func (s *apiV1) getAgreements(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
 	res := &rbody.TribeListAgreement{}
-	res.Agreements = s.tr.GetAgreements()
-	respond(200, res, w)
+	res.Agreements = s.tribeManager.GetAgreements()
+	rbody.Write(200, res, w)
 }
 
-func (s *Server) getAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) getAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	tribeLogger = tribeLogger.WithField("_block", "getAgreement")
 	name := p.ByName("name")
-	if _, ok := s.tr.GetAgreements()[name]; !ok {
+	if _, ok := s.tribeManager.GetAgreements()[name]; !ok {
 		fields := map[string]interface{}{
 			"agreement_name": name,
 		}
 		tribeLogger.WithFields(fields).Error(ErrAgreementDoesNotExist)
-		respond(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
+		rbody.Write(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
 		return
 	}
 	a := &rbody.TribeGetAgreement{}
 	var serr serror.SnapError
-	a.Agreement, serr = s.tr.GetAgreement(name)
+	a.Agreement, serr = s.tribeManager.GetAgreement(name)
 	if serr != nil {
 		tribeLogger.Error(serr)
-		respond(400, rbody.FromSnapError(serr), w)
+		rbody.Write(400, rbody.FromSnapError(serr), w)
 		return
 	}
-	respond(200, a, w)
+	rbody.Write(200, a, w)
 }
 
-func (s *Server) deleteAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) deleteAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	tribeLogger = tribeLogger.WithField("_block", "deleteAgreement")
 	name := p.ByName("name")
-	if _, ok := s.tr.GetAgreements()[name]; !ok {
+	if _, ok := s.tribeManager.GetAgreements()[name]; !ok {
 		fields := map[string]interface{}{
 			"agreement_name": name,
 		}
 		tribeLogger.WithFields(fields).Error(ErrAgreementDoesNotExist)
-		respond(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
+		rbody.Write(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
 		return
 	}
 
 	var serr serror.SnapError
-	serr = s.tr.RemoveAgreement(name)
+	serr = s.tribeManager.RemoveAgreement(name)
 	if serr != nil {
 		tribeLogger.Error(serr)
-		respond(400, rbody.FromSnapError(serr), w)
+		rbody.Write(400, rbody.FromSnapError(serr), w)
 		return
 	}
 
 	a := &rbody.TribeDeleteAgreement{}
-	a.Agreements = s.tr.GetAgreements()
-	respond(200, a, w)
+	a.Agreements = s.tribeManager.GetAgreements()
+	rbody.Write(200, a, w)
 }
 
-func (s *Server) joinAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) joinAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	tribeLogger = tribeLogger.WithField("_block", "joinAgreement")
 	name := p.ByName("name")
-	if _, ok := s.tr.GetAgreements()[name]; !ok {
+	if _, ok := s.tribeManager.GetAgreements()[name]; !ok {
 		fields := map[string]interface{}{
 			"agreement_name": name,
 		}
 		tribeLogger.WithFields(fields).Error(ErrAgreementDoesNotExist)
-		respond(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
+		rbody.Write(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
 		return
 	}
 
 	b, err := ioutil.ReadAll(r.Body)
 	if err != nil {
 		tribeLogger.Error(err)
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
 
@@ -125,37 +125,37 @@ func (s *Server) joinAgreement(w http.ResponseWriter, r *http.Request, p httprou
 		}
 		se := serror.New(ErrInvalidJSON, fields)
 		tribeLogger.WithFields(fields).Error(ErrInvalidJSON)
-		respond(400, rbody.FromSnapError(se), w)
+		rbody.Write(400, rbody.FromSnapError(se), w)
 		return
 	}
 
-	serr := s.tr.JoinAgreement(name, m.MemberName)
+	serr := s.tribeManager.JoinAgreement(name, m.MemberName)
 	if serr != nil {
 		tribeLogger.Error(serr)
-		respond(400, rbody.FromSnapError(serr), w)
+		rbody.Write(400, rbody.FromSnapError(serr), w)
 		return
 	}
-	agreement, _ := s.tr.GetAgreement(name)
-	respond(200, &rbody.TribeJoinAgreement{Agreement: agreement}, w)
+	agreement, _ := s.tribeManager.GetAgreement(name)
+	rbody.Write(200, &rbody.TribeJoinAgreement{Agreement: agreement}, w)
 
 }
 
-func (s *Server) leaveAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) leaveAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	tribeLogger = tribeLogger.WithField("_block", "leaveAgreement")
 	name := p.ByName("name")
-	if _, ok := s.tr.GetAgreements()[name]; !ok {
+	if _, ok := s.tribeManager.GetAgreements()[name]; !ok {
 		fields := map[string]interface{}{
 			"agreement_name": name,
 		}
 		tribeLogger.WithFields(fields).Error(ErrAgreementDoesNotExist)
-		respond(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
+		rbody.Write(400, rbody.FromSnapError(serror.New(ErrAgreementDoesNotExist, fields)), w)
 		return
 	}
 
 	b, err := ioutil.ReadAll(r.Body)
 	if err != nil {
 		tribeLogger.Error(err)
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
 
@@ -170,35 +170,35 @@ func (s *Server) leaveAgreement(w http.ResponseWriter, r *http.Request, p httpro
 		}
 		se := serror.New(ErrInvalidJSON, fields)
 		tribeLogger.WithFields(fields).Error(ErrInvalidJSON)
-		respond(400, rbody.FromSnapError(se), w)
+		rbody.Write(400, rbody.FromSnapError(se), w)
 		return
 	}
 
-	serr := s.tr.LeaveAgreement(name, m.MemberName)
+	serr := s.tribeManager.LeaveAgreement(name, m.MemberName)
 	if serr != nil {
 		tribeLogger.Error(serr)
-		respond(400, rbody.FromSnapError(serr), w)
+		rbody.Write(400, rbody.FromSnapError(serr), w)
 		return
 	}
-	agreement, _ := s.tr.GetAgreement(name)
-	respond(200, &rbody.TribeLeaveAgreement{Agreement: agreement}, w)
+	agreement, _ := s.tribeManager.GetAgreement(name)
+	rbody.Write(200, &rbody.TribeLeaveAgreement{Agreement: agreement}, w)
 }
 
-func (s *Server) getMembers(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
-	members := s.tr.GetMembers()
-	respond(200, &rbody.TribeMemberList{Members: members}, w)
+func (s *apiV1) getMembers(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+	members := s.tribeManager.GetMembers()
+	rbody.Write(200, &rbody.TribeMemberList{Members: members}, w)
 }
 
-func (s *Server) getMember(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) getMember(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	tribeLogger = tribeLogger.WithField("_block", "getMember")
 	name := p.ByName("name")
-	member := s.tr.GetMember(name)
+	member := s.tribeManager.GetMember(name)
 	if member == nil {
 		fields := map[string]interface{}{
 			"name": name,
 		}
 		tribeLogger.WithFields(fields).Error(ErrMemberNotFound)
-		respond(404, rbody.FromSnapError(serror.New(ErrMemberNotFound, fields)), w)
+		rbody.Write(404, rbody.FromSnapError(serror.New(ErrMemberNotFound, fields)), w)
 		return
 	}
 	resp := &rbody.TribeMemberShow{
@@ -213,15 +213,15 @@ func (s *Server) getMember(w http.ResponseWriter, r *http.Request, p httprouter.
 			resp.TaskAgreements = append(resp.TaskAgreements, k)
 		}
 	}
-	respond(200, resp, w)
+	rbody.Write(200, resp, w)
 }
 
-func (s *Server) addAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
+func (s *apiV1) addAgreement(w http.ResponseWriter, r *http.Request, p httprouter.Params) {
 	tribeLogger = tribeLogger.WithField("_block", "addAgreement")
 	b, err := ioutil.ReadAll(r.Body)
 	if err != nil {
 		tribeLogger.Error(err)
-		respond(500, rbody.FromError(err), w)
+		rbody.Write(500, rbody.FromError(err), w)
 		return
 	}
 
@@ -234,7 +234,7 @@ func (s *Server) addAgreement(w http.ResponseWriter, r *http.Request, p httprout
 		}
 		se := serror.New(ErrInvalidJSON, fields)
 		tribeLogger.WithFields(fields).Error(ErrInvalidJSON)
-		respond(400, rbody.FromSnapError(se), w)
+		rbody.Write(400, rbody.FromSnapError(se), w)
 		return
 	}
 
@@ -244,19 +244,19 @@ func (s *Server) addAgreement(w http.ResponseWriter, r *http.Request, p httprout
 		}
 		se := serror.New(ErrInvalidJSON, fields)
 		tribeLogger.WithFields(fields).Error(ErrInvalidJSON)
-		respond(400, rbody.FromSnapError(se), w)
+		rbody.Write(400, rbody.FromSnapError(se), w)
 		return
 	}
 
-	err = s.tr.AddAgreement(a.Name)
+	err = s.tribeManager.AddAgreement(a.Name)
 	if err != nil {
 		tribeLogger.WithField("agreement-name", a.Name).Error(err)
-		respond(400, rbody.FromError(err), w)
+		rbody.Write(400, rbody.FromError(err), w)
 		return
 	}
 
 	res := &rbody.TribeAddAgreement{}
-	res.Agreements = s.tr.GetAgreements()
+	res.Agreements = s.tribeManager.GetAgreements()
 
-	respond(200, res, w)
+	rbody.Write(200, res, w)
 }
